source/requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.2.5
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.4
pylint==2.5.3
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
